+++ CHANGELOG.md
@@ -18,6 +18,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Used python API of demucs instead of CLI API by
   @asaah18 [#119](https://github.com/asaah18/video-music-remover/pull/119)
 
+### Refactored
+
+- Refactored pre-processing and post-processing prints and logs into events by @asaah18
+  in [#120](https://github.com/asaah18/video-music-remover/pull/120)
+
 ## [1.0.0](https://github.com/asaah18/video-music-remover/releases/tag/1.0) - 2025-05-24
 
 ### Added