+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.1.0](https://github.com/asaah18/video-music-remover/releases/tag/1.1.0) - 2025-09-13
 
 ### Added